requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
pyodbc==5.0.1
python-dotenv==1.0.0
azure-servicebus==7.11.4
stripe==5.5.0
bcrypt==4.0.1
gunicorn==21.2.0